requirements.txt
functions-framework==3.*
openai>=1.0.0
python-dotenv==1.0.0
google-cloud-pubsub==2.18.4
flask>=3.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0
